Add route group
Moving everything except for the opengraph image into a route group. Had to update imports but nothing else is changed.

This breaks the intercepted rotues. Clicking the links on the photos seem to do nothing. It's possible I've misunderstood how to configure this, but I can't seem to get it to work with `(...)` or `(..)(..)` either.